% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{BA}
\alias{BA}
\title{Calculate batting average}
\usage{
BA(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
Find batting average for batters with more than zero at bats.
Required fields from the Batting table are; "AB", and "H."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- BA(batting_df)
new_df
}

}
\keyword{BA}
\keyword{ball}
\keyword{base}
\keyword{bb}
\keyword{on}
\keyword{percentage}

