% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{BABIP}
\alias{BABIP}
\title{Calculate batting average on balls in play (BABIP)}
\usage{
BABIP(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}
}
\description{
Find BABIP for batters with more than zero at bats.
Required fields from the Batting table are; "AB", "BB", "H", "HBP", "SF", "SH", "HR"  and "SO."
}
\examples{
\dontrun{
get_bbdb("Batting")
new_df <- BABIP(batting_df)
new_df
}

}
\keyword{BABIP}
\keyword{ball}
\keyword{base}
\keyword{bb}
\keyword{on}
\keyword{percentage}
