% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{BB_9}
\alias{BB_9}
\title{Calculate walks per nine innings}
\usage{
BB_9(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
Find batting average walks per nine innings for pitchers with more one or more inning pitched.
Required fields from the Batting table are; "IPouts", and "BB."
}
\examples{
\dontrun{
pitching_df <- Lahman::Pitching
new_df <- BB_9(pitching_df)
new_df
}

}
\keyword{BB}
\keyword{BB9}
\keyword{BB_9}
\keyword{bb/9}
