% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{BBpct}
\alias{BBpct}
\title{Calculate base on ball percentage}
\usage{
BBpct(dat, position)
}
\arguments{
\item{dat}{The data you would wish to calculate.}

\item{position}{This is a required argument that designates either a "batter" or "pitcher." The argument should
be spelled out and within quotes.}
}
\description{
Find base on ball percentage for batters or pitchers with more than zero at bats or innings pitched.
Required fields for batters are; "AB", "SO", "BB", "HBP", "SF", and "SH." Required fields for pitchers are;
"BB" and "IPouts". Intentional base on balls (IBB) is added for the years that metric is available.
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- BBpct(batting_df, position = "batter")
new_df
}

}
\keyword{BBpct}
\keyword{ball}
\keyword{base}
\keyword{bb}
\keyword{on}
\keyword{percentage}

