% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldingStats.R
\name{Ch}
\alias{Ch}
\title{Calculate defensive chances}
\usage{
Ch(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
The number of chances a player had to make a defensive play.
Required fields from the Fielding table are; "PO", "A", and "E."
}
\examples{
\dontrun{
get_bbdb("Fielding")
Fielding$Ch <- Ch(Fielding)

}

}
\keyword{Ch}
\keyword{Chances}
\keyword{Defensive}
