% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fip.R
\name{fip_values}
\alias{fip_values}
\title{Return fip values per season}
\usage{
fip_values(Fangraphs = FALSE)
}
\arguments{
\item{Fangraphs}{if TRUE the function will return the Fangraphs wOBA values. By default the function uses a method adapted from
Tom Tango. These values are often very close to Fangraphs, but aren't the same.}
}
\description{
Get fip values for each season.
}
\examples{
\dontrun{
fip_df <- fip_values(Fangraphs=FALSE)
head(fip_df)
}

}
\keyword{average,}
\keyword{base}
\keyword{fangraphs}
\keyword{on}
\keyword{wOBA,}
\keyword{woba,}

