% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldingStats.R
\name{Fld_pct}
\alias{Fld_pct}
\title{Calculate batting average}
\usage{
Fld_pct(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find batting average for batters with more than zero at bats.
Required fields from the Fielding table are; "PO", "A", and "E."
}
\examples{
\dontrun{
get_bbdb("Fielding")
Fielding$Fld_pct <- Fld_pct(Fielding)

}

}
\keyword{Fld_pct}
\keyword{fielding}
\keyword{percentage}
