% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{H_9}
\alias{H_9}
\title{Calculate Hits per Nine innings}
\usage{
H_9(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
Find the number of hits a pitcher throws per nine innings pitched.
Required fields from the Batting table are; "H" and "IPouts."
}
\examples{
\dontrun{
pitching_df <- Lahman::Pitching
new_df <- H_9(pitching_df)
new_df
}

}
\keyword{hits}
\keyword{innings}
\keyword{nine}
\keyword{per}
