% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{Kpct}
\alias{Kpct}
\title{Calculate strikeout percentage}
\usage{
Kpct(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}
}
\description{
Find strikeout percentage for batters with more than zero at bats.
Required fields from the Batting table are; "AB", "SO", "BB", "HBP", "SF", and "SH."
}
\examples{
\dontrun{
get_bbdb("Batting")
Batting$Kpct <- Kpct(Batting)

}

}
\keyword{Kpct}
\keyword{percentage}
\keyword{strikeout}
