% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{LOB_pct}
\alias{LOB_pct}
\title{Calculate the left on base percentage}
\usage{
LOB_pct(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
Find the percentaqge of base runners that a pitcher leaves on base of the course of a season.
Required fields from the Batting table are; "H", "BB", "HBP", "R", and "HR."
}
\examples{
\dontrun{
pitching_df <- Lahman::Pitching
new_df <- LOB_pct(pitching_df)
new_df
}

}
\keyword{LOB}
\keyword{LOB_pct}
\keyword{percentage}
