% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{PA}
\alias{PA}
\title{Calculate plate appearances for batters}
\usage{
PA(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
Find the plate appearances (PA) for batters.
Required fields from the batting table are "AB", "BB", "HBP", "SH", and "SF."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- PA(batting_df)
new_df
}

}
\keyword{PA}
\keyword{base}
\keyword{on}
\keyword{percentage}

