% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching2016.R
\docType{data}
\name{Pitching2016}
\alias{Pitching2016}
\title{A sample subset of the Pitching table from the Baseball Databank for the year 2016.}
\format{A data frame with 824 rows and 30 variables:
\describe{
  \item{playerID}{database key for unique player}
  \item{yearID}{year}
  \item{stint}{number of times played on team in a single year}
  \item{teamID}{database key for unique team}
  \item{lgID}{database key for unique league}
  \item{W}{number of games won}
  \item{L}{number of games lost}
  \item{G}{number of games played}
  \item{GS}{number of games started}
  \item{CG}{number of complete games}
  \item{SHO}{number of shutouts}
  \item{SV}{number of home saves}
  \item{IPouts}{number of outs pitched}
  \item{H}{number of hits surrendered}
  \item{ER}{number of earned runs}
  \item{HR}{number of home runs surrendered}
  \item{BB}{number of base on balls}
  \item{SO}{number of strike outs}
  \item{BAOpp}{opponent batting average}
  \item{ERA}{earned run average}
  \item{IBB}{number of intentional base on balls}
  \item{WP}{number of wild pitches}
  \item{HBP}{number of hit batters by pitch}
  \item{BK}{number of balks}
  \item{BFP}{batters faced by pitcher}
  \item{GF}{Games finished}
  \item{R}{number of runs allowed}
  \item{SH}{number of sacrifice hits by opposing batters}
  \item{SF}{number of sacrifice flys by opposing batters}
  \item{GIDP}{number of grounded into double plays by opposing batters}

}}
\usage{
data(Pitching2016)
}
\description{
A dataset containing Pitching statistics in 2016.
}
\note{
Last updated 2016-06-15
}
\keyword{internal}
