% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{RCtech}
\alias{RCtech}
\title{Calculate Runs Created using the technical formula.}
\usage{
RCtech(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
The "Technical Version" is the most well-known formula for RC. It adds several factors to the
basic formula such as sacrifice hits, stolen bases and intentional base on balls.
Required fields from the batting table are "AB", "H", "BB", "X2B", "X3B", "HR", "GIDP", "HBP", "SB", "CS",
"SF" and "SH," and "IBB."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- RCtech(batting_df)
new_df
}

}
\keyword{RCtech}
\keyword{base}
\keyword{extra}
\keyword{hit}
\keyword{per}
