% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{SLG}
\alias{SLG}
\title{Calculate slugging percentage (SLG)}
\usage{
SLG(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find the SLG for batters with more than zero hits.
Required fields from the batting table are "H", "X2B", "X3B", "HR"."
}
\examples{
\dontrun{
get_bbdb("Batting")
Batting$SLG <- SLG(Batting)

}

}
\keyword{SLG}
\keyword{base}
\keyword{on}
\keyword{percentage}
