% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{TBs}
\alias{TBs}
\title{Calculate a batter's total bases}
\usage{
TBs(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
Find total bases.
Required fields from the batting table are "AB","H", "X2B", "X3B" and "HR."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- TBs(batting_df)
new_df
}

}
\keyword{TBs}
\keyword{bases}
\keyword{total}
