% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{WHIP}
\alias{WHIP}
\title{Calculate Walks plus Hits per Innings Pitched}
\usage{
WHIP(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
Find the number of walks plus hits a pitcher allows per inning pitched.
Required fields from the Batting table are; "H", "BB", and "IPouts."
}
\examples{
\dontrun{
pitching_df <- Lahman::Pitching
new_df <- WHIP(pitching_df)
new_df
}

}
\keyword{Hits}
\keyword{Innings}
\keyword{Pitched}
\keyword{WHIP}
\keyword{Walks}
\keyword{per}
\keyword{plus}
