% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWOBA.R
\name{wOBA}
\alias{wOBA}
\title{Return wOBA values}
\usage{
wOBA()
}
\description{
Get wOBA values for each year in your database. This calculation requires all fields of
the Pitching, Fielding and Batting tables from the Lahman package.
}
\examples{

woba_df <- wOBA()
head(woba_df)


}
\keyword{average}
\keyword{base}
\keyword{on}
\keyword{wOBA,}
\keyword{woba,}

