% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woba.R
\name{wOBA_values}
\alias{wOBA_values}
\title{Return wOBA values per season}
\usage{
wOBA_values(Sep.Leagues = FALSE, Fangraphs = FALSE)
}
\arguments{
\item{Sep.Leagues}{If TRUE, this will split the calculation and return unique wOBA values for the various leagues. This can be
helpful in handling Designated Hitters and National League pitchers. It also isolates the park factors to their respective leagues.}

\item{Fangraphs}{if TRUE the function will return the Fangraphs wOBA values. By default the function uses a method adapted from
Tom Tango. These values are often very close to Fangraphs, but are not the same due to Fangraphs using a different algorithum.
This can not be used in conjuction with the \code{Sep.Leagues} argument because Fangraphs does not seperate FIP constants by league.}
}
\description{
Get wOBA values for each year in your database. This calculation requires all fields of
the Pitching, Fielding and Batting tables from the Lahman package, or a comprable data set. The function uses
a version of Tom Tango's wOBA formula by default, but can also return Fangraphs wOBA values.
}
\examples{
\dontrun{
woba_df <- wOBA_values(Sep.Leagues=FALSE, Fangraphs=FALSE)
head(woba_df)
}

}
\keyword{average,}
\keyword{base}
\keyword{fangraphs}
\keyword{on}
\keyword{wOBA,}
\keyword{woba,}

