% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{wRAA}
\alias{wRAA}
\title{Calculate Weighted Runs Above Average (wRAA)}
\usage{
wRAA(dat = NULL, Fangraphs = FALSE, NA_to_zero = TRUE,
  Sep.Leagues = FALSE)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}

\item{Fangraphs}{If TRUE the function will download wOBA values from Fangraphs. Both wOBA scale and leauge wOBA are used in the wRAA
calculation. If FALSE the function will use the internal wOBA algorithum, which is adapted from Tom Tango's original wOBA formula.
This algorithum produces a slightly different wOBA scale than the Fangraphs wOBA scale, so variations in wRAA should be expected.
The default internal method does not require an external download from Fangraphs. If not specified, the default is set to FALSE.}

\item{NA_to_zero}{If TRUE this will replace NAs with 0 for years that certain stats weren't counted. For example, sacrafice hits
weren't a counted statistic until 1954, therefore we are technically unable to calculate wRAA for any player prior to 1954.
The default is set to TRUE. Even though this is bad practice mathematically, many in the sabermetrics community accept the practice.
If FALSE, the wRAA calculation will return NaN for years with missing data.}

\item{Sep.Leagues}{If TRUE the algorithum will calculate different run enviornments for the National and American leagues. Grouping
the leauges can solve problems introduced by the designated hitter and hitting pitchers. It also serves to further isolate for
park factors between the American and National leauges. The default for this argument is FALSE.}
}
\description{
Find the wRAA for all players with one or more hits for a particular season.
Required fields from the batting table are "AB", "H", "BB", "X2B", "X3B", "HR", "HBP", "SF", "IBB."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- wRAA(batting_df, Fangraphs=FALSE, NA_to_zero=TRUE, Sep.Leagues=FALSE)
new_df
}

}
\keyword{Above}
\keyword{Average}
\keyword{Runs}
\keyword{Weighted}
\keyword{wRAA}

