% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{BA}
\alias{BA}
\title{Batting: Calculate batting average}
\usage{
BA(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find batting average for batters with more than zero at bats.
Required fields from the Batting table are; "AB", and "H."
}
\examples{

data("Batting2016")
head(Batting2016)

Batting2016$BA <- BA(Batting2016)

}
\seealso{
Other Batting functions: \code{\link{BABIP}},
  \code{\link{BBpct}}, \code{\link{CTpct}},
  \code{\link{HRpct}}, \code{\link{ISO}},
  \code{\link{Kpct}}, \code{\link{OBP}}, \code{\link{OPS}},
  \code{\link{PA}}, \code{\link{RC2002}},
  \code{\link{RCbasic}}, \code{\link{RCtech}},
  \code{\link{SLG}}, \code{\link{TBs}},
  \code{\link{XBHpct}}, \code{\link{XBperH}},
  \code{\link{wOBA}}, \code{\link{wRAA}}, \code{\link{wRC}}
}
\keyword{BA}
\keyword{ball}
\keyword{base}
\keyword{bb}
\keyword{on}
\keyword{percentage}
