% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{BABIP}
\alias{BABIP}
\title{Calculate batting average on balls in play (BABIP)}
\usage{
BABIP(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
Find BABIP for batters with more than zero at bats.
Required fields from the Batting table are; "AB", "BB", "H", "HBP", "SF", "SH", "HR"  and "SO."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- BABIP(batting_df)
new_df
}

}
\keyword{BABIP}
\keyword{ball}
\keyword{base}
\keyword{bb}
\keyword{on}
\keyword{percentage}

