% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{BABIP}
\alias{BABIP}
\title{Batting: Calculate batting average on balls in play (BABIP)}
\usage{
BABIP(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find BABIP for batters with more than zero at bats.
Required fields from the Batting table are; "AB", "BB", "H", "HBP", "SF", "SH", "HR"  and "SO."
}
\examples{

data("Batting2016")
head(Batting2016)

Batting2016$BABIP <- BABIP(Batting2016)

}
\seealso{
Other Batting functions: \code{\link{BA}},
  \code{\link{BBpct}}, \code{\link{CTpct}},
  \code{\link{HRpct}}, \code{\link{ISO}},
  \code{\link{Kpct}}, \code{\link{OBP}}, \code{\link{OPS}},
  \code{\link{PA}}, \code{\link{RC2002}},
  \code{\link{RCbasic}}, \code{\link{RCtech}},
  \code{\link{SLG}}, \code{\link{TBs}},
  \code{\link{XBHpct}}, \code{\link{XBperH}},
  \code{\link{wOBA}}, \code{\link{wRAA}}, \code{\link{wRC}}
}
\keyword{BABIP}
\keyword{ball}
\keyword{base}
\keyword{bb}
\keyword{on}
\keyword{percentage}
