% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseballDBR.R
\name{baseballDBR}
\alias{baseballDBR}
\title{baseballDBR: A package for working with data from the Baseball Databank/Lahman Database.}
\description{
baseballDBR: A package for working with data from the Baseball Databank/Lahman Database.
}
