% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{BB_9}
\alias{BB_9}
\title{Calculate walks per nine innings}
\usage{
BB_9(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}
}
\description{
Find batting average walks per nine innings for pitchers with more one or more inning pitched.
Required fields from the Pitching table are; "IPouts", and "BB."
}
\examples{
\dontrun{
get_bbdb("Pitching")
new_df <- BB_9(Pitching)
new_df
}

}
\keyword{BB}
\keyword{BB9}
\keyword{BB_9}
\keyword{bb/9}
