% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{BB_9}
\alias{BB_9}
\title{Pitching: Calculate walks per nine innings}
\usage{
BB_9(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find batting average walks per nine innings for pitchers with more one or more inning pitched.
Required fields from the Pitching table are; "IPouts", and "BB."
}
\examples{

data("Pitching2016")
head(Pitching2016)

Pitching2016$BB_9 <- BB_9(Pitching2016)

}
\seealso{
Other Pitching functions: \code{\link{FIP}},
  \code{\link{HR_9}}, \code{\link{H_9}}, \code{\link{IP}},
  \code{\link{K_9}}, \code{\link{LOB_pct}},
  \code{\link{WHIP}}
}
\keyword{BB}
\keyword{BB9}
\keyword{BB_9}
\keyword{bb/9}
