% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{BBpct}
\alias{BBpct}
\title{Calculate base on ball percentage}
\usage{
BBpct(dat)
}
\arguments{
\item{dat}{The data you would wish to calculate.}
}
\description{
Find base on ball percentage for batters with more than zero at bats.
Required fields from the Batting table are; "AB", "SO", "BB", "HBP", "SF", and "SH."
Intentional base on balls (IBB) is added for the years that metric is available.
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- BBpct(batting_df)
new_df
}

}
\keyword{BBpct}
\keyword{ball}
\keyword{base}
\keyword{bb}
\keyword{on}
\keyword{percentage}

