% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{BBpct}
\alias{BBpct}
\title{Calculate base on ball percentage}
\usage{
BBpct(dat)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}
}
\description{
Find base on ball percentage for batters with more than zero at bats.
Required fields from the Batting table are; "AB", "SO", "BB", "HBP", "SF", and "SH."
Intentional base on balls (IBB) is added for the years that metric is available.
}
\examples{
\dontrun{
get_bbdb("Batting")
Batting$BBpct <- BBpct(Batting)

}

}
\keyword{BBpct}
\keyword{ball}
\keyword{base}
\keyword{bb}
\keyword{on}
\keyword{percentage}
