% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{CTpct}
\alias{CTpct}
\title{Calculate a batter's contact rate}
\usage{
CTpct(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}
}
\description{
Find the contact rate for batters.
Required fields from the batting table are "AB" and "SO."
}
\examples{
\dontrun{
get_bbdb("Batting")
Batting$CTpct <- CTpct(Batting)

}

}
\keyword{CTpct}
\keyword{contact}
\keyword{rate}
