% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{CTpct}
\alias{CTpct}
\title{Calculate a batter's contact rate}
\usage{
CTpct(dat)
}
\arguments{
\item{dat}{The data you would wish to calculate.}
}
\description{
Find the contact rate for batters.
Required fields from the batting table are "AB" and "SO."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- CTpct(batting_df)
new_df
}

}
\keyword{CTpct}
\keyword{contact}
\keyword{rate}

