% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{H_9}
\alias{H_9}
\title{Calculate Hits per Nine innings}
\usage{
H_9(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate.  The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}
}
\description{
Find the number of hits a pitcher throws per nine innings pitched.
Required fields from the Pitching table are; "H", "BB", and "IPouts."
}
\examples{
\dontrun{
get_bbdb("Pitching")
Pitching$H_9 <- H_9(Pitching)

}

}
\keyword{hits}
\keyword{innings}
\keyword{nine}
\keyword{per}
