% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{H_9}
\alias{H_9}
\title{Pitching: Calculate Hits per Nine innings}
\usage{
H_9(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate.  The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find the number of hits a pitcher throws per nine innings pitched.
Required fields from the Pitching table are; "H", "BB", and "IPouts."
}
\examples{

data("Pitching2016")
head(Pitching2016)

Pitching2016$H_9 <- H_9(Pitching2016)

}
\seealso{
Other Pitching functions: \code{\link{BB_9}},
  \code{\link{FIP}}, \code{\link{HR_9}}, \code{\link{IP}},
  \code{\link{K_9}}, \code{\link{LOB_pct}},
  \code{\link{WHIP}}
}
\keyword{hits}
\keyword{innings}
\keyword{nine}
\keyword{per}
