% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{HR_9}
\alias{HR_9}
\title{Pitching: Calculate Home Runs per Nine innings}
\usage{
HR_9(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find the number of home runs a pitcher allows per nine innings pitched.
Required fields from the Pitching table are; "H" and "IPouts."
}
\examples{

data("Pitching2016")
head(Pitching2016)

Pitching2016$HR_9 <- HR_9(Pitching2016)

}
\seealso{
Other Pitching functions: \code{\link{BB_9}},
  \code{\link{FIP}}, \code{\link{H_9}}, \code{\link{IP}},
  \code{\link{K_9}}, \code{\link{LOB_pct}},
  \code{\link{WHIP}}
}
\keyword{hits}
\keyword{innings}
\keyword{nine}
\keyword{per}
