% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{HRpct}
\alias{HRpct}
\title{Calculate home run percentage}
\usage{
HRpct(dat)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}
}
\description{
Find home run percentage for batters with more than zero at bats.
Required fields from the Batting table are "AB" and "HR."
}
\examples{
\dontrun{
get_bbdb("Batting")
Batting$HRpct <- HRpct(Batting)

}

}
\keyword{HRpct}
\keyword{home}
\keyword{percentage}
\keyword{run}
