% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{HRpct}
\alias{HRpct}
\title{Calculate home run percentage}
\usage{
HRpct(dat)
}
\arguments{
\item{dat}{The data you would wish to calculate.}
}
\description{
Find home run percentage for batters with more than zero at bats.
Required fields from the Batting table are "AB" and "HR."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- HRpct(batting_df)
new_df
}

}
\keyword{HRpct}
\keyword{home}
\keyword{percentage}
\keyword{run}

