% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{IP}
\alias{IP}
\title{Calculate the innings pitched}
\usage{
IP(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
Find the number of innings a player has ptiched for a season.
Required fields from the Batting table are; "IPouts."
}
\examples{
\dontrun{
pitching_df <- Lahman::Pitching
new_df <- IP(pitching_df)
new_df
}

}
\keyword{innings}
\keyword{pitched}
