% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{IP}
\alias{IP}
\title{Calculate the innings pitched}
\usage{
IP(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}
}
\description{
Find the number of innings a player has ptiched for a season.
Required fields from the Pitching table are; "IPouts."
}
\examples{
\dontrun{
get_bbdb("Pitching")
new_df <- IP(Pitching)
new_df
}

}
\keyword{innings}
\keyword{pitched}
