% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{IP}
\alias{IP}
\title{Pitching: Calculate the innings pitched}
\usage{
IP(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find the number of innings a player has pitched for a season.
Required fields from the Pitching table are; "IPouts."
}
\examples{

data("Pitching2016")
head(Pitching2016)

Pitching2016$IP <- IP(Pitching2016)

}
\seealso{
Other Pitching functions: \code{\link{BB_9}},
  \code{\link{FIP}}, \code{\link{HR_9}}, \code{\link{H_9}},
  \code{\link{K_9}}, \code{\link{LOB_pct}},
  \code{\link{WHIP}}
}
\keyword{innings}
\keyword{pitched}
