% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{ISO}
\alias{ISO}
\title{Calculate ISO for batters}
\usage{
ISO(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
Find isolated power (ISO) for batters with more than zero at bats.
Required fields from the batting table are "H", "X2B", "X3B", "HR"."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- ISO(batting_df)
new_df
}

}
\keyword{ISO}
\keyword{isolated}
\keyword{power}
