% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{ISO}
\alias{ISO}
\title{Calculate ISO for batters}
\usage{
ISO(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find isolated power (ISO) for batters with more than zero at bats.
Required fields from the batting table are "H", "X2B", "X3B", "HR"."
}
\examples{
\dontrun{
get_bbdb("Batting")
Batting$ISO <- ISO(Batting)

}

}
\keyword{ISO}
\keyword{isolated}
\keyword{power}
