% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{K_9}
\alias{K_9}
\title{Calculate Strikes per Nine innings}
\usage{
K_9(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}
}
\description{
Find the number of strikes a pitcher throws per nine innings pitched.
Required fields from the Pitching table are; "H", "BB", "IPouts", and "SO."
}
\examples{
\dontrun{
get_bbdb("Pitching")
Pitching$K_9 <- K_9(Pitching)

}

}
\keyword{innings}
\keyword{nine}
\keyword{per}
\keyword{strikes}
