% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{Kpct}
\alias{Kpct}
\title{Calculate strikeout percentage}
\usage{
Kpct(dat)
}
\arguments{
\item{dat}{The data you would wish to calculate.}
}
\description{
Find strikeout percentage for batters with more than zero at bats.
Required fields from the Batting table are; "AB", "SO", "BB", "HBP", "SF", and "SH."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- Kpct(batting_df)
new_df
}

}
\keyword{Kpct}
\keyword{percentage}
\keyword{strikeout}

