% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{Kpct}
\alias{Kpct}
\title{Calculate strikeout percentage}
\usage{
Kpct(dat, position)
}
\arguments{
\item{dat}{The data you would wish to calculate.}

\item{position}{This is a required argument that designates either a "batter" or "pitcher." The argument should
be spelled out and within quotes.}
}
\description{
Find strikeout percentage for batters or pitchers with more than zero at bats or innings pitched.
Required fields for batters are; "AB", "SO", "BB", "HBP", "SF", and "SH." Required fields for pitchers are;
"SO" and "IPouts"
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- Kpct(batting_df, position = "batter")
new_df
}

}
\keyword{Kpct}
\keyword{percentage}
\keyword{strikeout}

