% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{Kpct}
\alias{Kpct}
\title{Batting: Calculate strikeout percentage}
\usage{
Kpct(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find strikeout percentage for batters with more than zero at bats.
Required fields from the Batting table are; "AB", "SO", "BB", "HBP", "SF", and "SH."
}
\examples{

data("Batting2016")
head(Batting2016)

Batting2016$Kpct <- Kpct(Batting2016)

}
\seealso{
Other Batting functions: \code{\link{BABIP}},
  \code{\link{BA}}, \code{\link{BBpct}},
  \code{\link{CTpct}}, \code{\link{HRpct}},
  \code{\link{ISO}}, \code{\link{OBP}}, \code{\link{OPS}},
  \code{\link{PA}}, \code{\link{RC2002}},
  \code{\link{RCbasic}}, \code{\link{RCtech}},
  \code{\link{SLG}}, \code{\link{TBs}},
  \code{\link{XBHpct}}, \code{\link{XBperH}},
  \code{\link{wOBA}}, \code{\link{wRAA}}, \code{\link{wRC}}
}
\keyword{Kpct}
\keyword{percentage}
\keyword{strikeout}
