% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{LOB_pct}
\alias{LOB_pct}
\title{Calculate the left on base percentage}
\usage{
LOB_pct(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}
}
\description{
Find the percentaqge of base runners that a pitcher leaves on base of the course of a season.
Required fields from the Pitching table are; "H", "BB", "HBP", "R", and "HR."
}
\examples{
\dontrun{
get_bbdb("Pitching")
Pitching$LOB_pct <- LOB_pct(pitching_df)

}

}
\keyword{LOB}
\keyword{LOB_pct}
\keyword{percentage}
