% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{LOB_pct}
\alias{LOB_pct}
\title{Pitching: Calculate the left on base percentage}
\usage{
LOB_pct(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find the percentage of base runners that a pitcher leaves on base of the course of a season.
Required fields from the Pitching table are; "H", "BB", "HBP", "R", and "HR."
}
\examples{

data("Pitching2016")
head(Pitching2016)

Pitching2016$LOB_pct <- LOB_pct(Pitching2016)

}
\seealso{
Other Pitching functions: \code{\link{BB_9}},
  \code{\link{FIP}}, \code{\link{HR_9}}, \code{\link{H_9}},
  \code{\link{IP}}, \code{\link{K_9}}, \code{\link{WHIP}}
}
\keyword{LOB}
\keyword{LOB_pct}
\keyword{percentage}
