% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{OBP}
\alias{OBP}
\title{Calculate on base percentage (OBP)}
\usage{
OBP(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
Find the OBP for batters with more than zero hits.
Required fields from the batting table are "H", "X2B", "X3B", "HR"."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- OBP(batting_df)
new_df
}

}
\keyword{OBP}
\keyword{base}
\keyword{on}
\keyword{percentage}

