% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{OPS}
\alias{OPS}
\title{Calculate on base percentage plus slugging (OPS)}
\usage{
OPS(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
Find the OPS for batters with more than zero hits.
Required fields from the batting table are "H", "X2B", "X3B", "HR", "BB", "HBP", "AB" and "SF."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- OPS(batting_df)
new_df
}

}
\keyword{OPS}
\keyword{base}
\keyword{on}
\keyword{percentage}
