% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{OPS}
\alias{OPS}
\title{Batting: Calculate on base percentage plus slugging (OPS)}
\usage{
OPS(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find the OPS for batters with more than zero hits.
Required fields from the batting table are "H", "X2B", "X3B", "HR", "BB", "HBP", "AB" and "SF."
}
\examples{

data("Batting2016")
head(Batting2016)

Batting2016$OPS <- OPS(Batting2016)

}
\seealso{
Other Batting functions: \code{\link{BABIP}},
  \code{\link{BA}}, \code{\link{BBpct}},
  \code{\link{CTpct}}, \code{\link{HRpct}},
  \code{\link{ISO}}, \code{\link{Kpct}}, \code{\link{OBP}},
  \code{\link{PA}}, \code{\link{RC2002}},
  \code{\link{RCbasic}}, \code{\link{RCtech}},
  \code{\link{SLG}}, \code{\link{TBs}},
  \code{\link{XBHpct}}, \code{\link{XBperH}},
  \code{\link{wOBA}}, \code{\link{wRAA}}, \code{\link{wRC}}
}
\keyword{OPS}
\keyword{base}
\keyword{on}
\keyword{percentage}
