% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{RC2002}
\alias{RC2002}
\title{Calculate Runs Created using the updated 2002 formula.}
\usage{
RC2002(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
The "2002 Version" is an updated version of the "Technical Version" by Bill James.
The 2002 RC uses the same counting stats as the Technical Version but applies weights to many of the raw stats.
Required fields from the batting table are "AB", "H", "BB", "X2B", "X3B", "HR", "GIDP", "HBP", "SB", "CS",
"SF" and "SH," "SO", and "IBB."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- RC2002(batting_df)
new_df
}

}
\keyword{RC2002}
\keyword{base}
\keyword{extra}
\keyword{hit}
\keyword{per}

