% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{RCbasic}
\alias{RCbasic}
\title{Calculate Runs Created using the basic formula.}
\usage{
RCbasic(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
Find the runs created using the basic formula presented by Bill James in the late 1970s.
Required fields from the batting table are "AB", "H", "BB", "X2B", "X3B", and "HR."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- RCbasic(batting_df)
new_df
}

}
\keyword{RCbasic}
\keyword{base}
\keyword{extra}
\keyword{hit}
\keyword{per}
