% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{RCtech}
\alias{RCtech}
\title{Calculate Runs Created using the technical formula.}
\usage{
RCtech(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}
}
\description{
The "Technical Version" is the most well-known formula for RC. It adds several factors to the
basic formula such as sacrifice hits, stolen bases and intentional base on balls.
Required fields from the batting table are "AB", "H", "BB", "X2B", "X3B", "HR", "GIDP", "HBP", "SB", "CS",
"SF" and "SH," and "IBB."
}
\examples{
\dontrun{
get_bbdb("Batting")
Batting$RCtech <- RCtech(Batting)

}

}
\keyword{RCtech}
\keyword{base}
\keyword{extra}
\keyword{hit}
\keyword{per}
