% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{SLG}
\alias{SLG}
\title{Batting: Calculate slugging percentage (SLG)}
\usage{
SLG(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find the SLG for batters with more than zero hits.
Required fields from the batting table are "H", "X2B", "X3B", "HR"."
}
\examples{

data("Batting2016")
head(Batting2016)

Batting2016$SLG <- SLG(Batting2016)

}
\seealso{
Other Batting functions: \code{\link{BABIP}},
  \code{\link{BA}}, \code{\link{BBpct}},
  \code{\link{CTpct}}, \code{\link{HRpct}},
  \code{\link{ISO}}, \code{\link{Kpct}}, \code{\link{OBP}},
  \code{\link{OPS}}, \code{\link{PA}},
  \code{\link{RC2002}}, \code{\link{RCbasic}},
  \code{\link{RCtech}}, \code{\link{TBs}},
  \code{\link{XBHpct}}, \code{\link{XBperH}},
  \code{\link{wOBA}}, \code{\link{wRAA}}, \code{\link{wRC}}
}
\keyword{SLG}
\keyword{base}
\keyword{on}
\keyword{percentage}
