% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{TBs}
\alias{TBs}
\title{Calculate a batter's total bases}
\usage{
TBs(dat)
}
\arguments{
\item{dat}{The data you would wish to calculate.}
}
\description{
Find total bases.
Required fields from the batting table are "H", "X2B", "X3B" and "HR."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- TBs(batting_df)
new_df
}

}
\keyword{TBs}
\keyword{bases}
\keyword{total}

