% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{TBs}
\alias{TBs}
\title{Calculate a batter's total bases}
\usage{
TBs(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find total bases.
Required fields from the batting table are "AB","H", "X2B", "X3B" and "HR."
}
\examples{
\dontrun{
get_bbdb("Batting")
Batting$TBs <- TBs(Batting)

}

}
\keyword{TBs}
\keyword{bases}
\keyword{total}
