% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{TBs}
\alias{TBs}
\title{Batting: Calculate a batter's total bases}
\usage{
TBs(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find total bases.
Required fields from the batting table are "AB","H", "X2B", "X3B" and "HR."
}
\examples{

data("Batting2016")
head(Batting2016)

Batting2016$TBs <- TBs(Batting2016)

}
\seealso{
Other Batting functions: \code{\link{BABIP}},
  \code{\link{BA}}, \code{\link{BBpct}},
  \code{\link{CTpct}}, \code{\link{HRpct}},
  \code{\link{ISO}}, \code{\link{Kpct}}, \code{\link{OBP}},
  \code{\link{OPS}}, \code{\link{PA}},
  \code{\link{RC2002}}, \code{\link{RCbasic}},
  \code{\link{RCtech}}, \code{\link{SLG}},
  \code{\link{XBHpct}}, \code{\link{XBperH}},
  \code{\link{wOBA}}, \code{\link{wRAA}}, \code{\link{wRC}}
}
\keyword{TBs}
\keyword{bases}
\keyword{total}
