% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{WHIP}
\alias{WHIP}
\title{Calculate Walks plus Hits per Innings Pitched}
\usage{
WHIP(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}
}
\description{
Find the number of walks plus hits a pitcher allows per inning pitched.
Required fields from the Pitching table are; "H", "BB", and "IPouts."
}
\examples{
\dontrun{
get_bbdb("Pitching")
new_df <- WHIP(pitching_df)
new_df
}

}
\keyword{Hits}
\keyword{Innings}
\keyword{Pitched}
\keyword{WHIP}
\keyword{Walks}
\keyword{per}
\keyword{plus}
