% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchingStats.R
\name{WHIP}
\alias{WHIP}
\title{Pitching: Calculate Walks plus Hits per Innings Pitched}
\usage{
WHIP(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.}
}
\description{
Find the number of walks plus hits a pitcher allows per inning pitched.
Required fields from the Pitching table are; "H", "BB", and "IPouts."
}
\examples{

data("Pitching2016")
head(Pitching2016)

Pitching2016$WHIP <- WHIP(Pitching2016)

}
\seealso{
Other Pitching functions: \code{\link{BB_9}},
  \code{\link{FIP}}, \code{\link{HR_9}}, \code{\link{H_9}},
  \code{\link{IP}}, \code{\link{K_9}},
  \code{\link{LOB_pct}}
}
\keyword{Hits}
\keyword{Innings}
\keyword{Pitched}
\keyword{WHIP}
\keyword{Walks}
\keyword{per}
\keyword{plus}
