% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{wOBA}
\alias{wOBA}
\title{Calculate Weighted On-Base Average (wOBA)}
\usage{
wOBA(dat = NULL, Fangraphs = FALSE, NA_to_zero = TRUE,
  Sep.Leagues = FALSE)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}

\item{Fangraphs}{If TRUE the function will download wOBA values from Fangraphs. If FALSE the function will use the internal
formula adapted from Tom Tango's original wOBA formula. Note, the internal formula is typically identical to Fangraphs and
does not require an external download. If not specified, the default is set to FALSE.}

\item{NA_to_zero}{If TRUE this will replace NAs with 0 for years that certain stats weren't counted. For example, sacrafice hits
weren't a counted statistic until 1954, therefore we are technically unable to calculate wOBA for any player prior to 1954.
The default is set to TRUE. Even though this is bad practice mathematically, many in the sabermetrics community accept the practice.
If FALSE, the wOBA calculation will return NaN for years with missing data.}

\item{Sep.Leagues}{If TRUE the algorithum will calculate different run enviornments for the National and American leagues. Grouping
the leauges can solve problems introduced by the designated hitter and hitting pitchers. It also serves to further isolate for
park factors between the American and National leauges. The default for this argument is FALSE.}
}
\description{
Find the wOBA for all players with one or more hits for a particular season.
Required fields from the batting table are "AB", "H", "BB", "X2B", "X3B", "HR", "HBP", "SF", "IBB."
}
\examples{
\dontrun{
get_bbdb("Batting")
new_df <- wOBA(Batting, Fangraphs=FALSE, NA_to_zero=TRUE, Sep.Leagues=FALSE)
new_df
}

}
\keyword{Average}
\keyword{On-Base}
\keyword{Weighted}
\keyword{wOBA}
