% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{XBHpct}
\alias{XBHpct}
\title{Calculate extra base percentage}
\usage{
XBHpct(dat)
}
\arguments{
\item{dat}{The data you would wish to calculate.}
}
\description{
Find extra base percentage for batters with more than zero at bats.
Required fields from the batting table are "AB", "BB", "HBP", "SF", "SH", "X2B", "X3B", "HR"."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- XBHpct(batting_df)
new_df
}

}
\keyword{XBHpct}
\keyword{base}
\keyword{extra}
\keyword{percentage}

