% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{XBHpct}
\alias{XBHpct}
\title{Calculate extra base percentage}
\usage{
XBHpct(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}
}
\description{
Find extra base percentage for batters with more than zero at bats.
Required fields from the batting table are "AB", "BB", "HBP", "SF", "SH", "X2B", "X3B", "HR"."
}
\examples{
\dontrun{
get_bbdb("Batting")
new_df <- XBHpct(Batting)
new_df
}

}
\keyword{XBHpct}
\keyword{base}
\keyword{extra}
\keyword{percentage}
