% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{XBHpct}
\alias{XBHpct}
\title{Calculate xtra base percentage}
\usage{
XBHpct(dat, position)
}
\arguments{
\item{dat}{The data you would wish to calculate.}

\item{position}{This is a required argument that designates either a "batter" or "pitcher." The argument should
be spelled out and within quotes.}
}
\description{
Find extra base percentage for batters or pitchers with more than zero at bats or innings pitched.
Required fields for batters are "AB", "BB", "HBP", "SF", "SH", "X2B", "X3B", "HR"."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- XBHpct(batting_df, position = "batter")
new_df
}

}
\keyword{XBHpct}
\keyword{base}
\keyword{percentage}
\keyword{xtra}

