% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{XBperH}
\alias{XBperH}
\title{Calculate extra base per hit}
\usage{
XBperH(dat)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. The data frame must have the same column names found in
The \code{Lahman} package or the Chadwick Bureau GitHub repository.
For a list of column names, use the \code{Lahman_names()} function.}
}
\description{
Find the average extra bases per hit for batters with more than zero hits.
Required fields from the batting table are "H", "X2B", "X3B", "HR"."
}
\examples{
\dontrun{
get_bbdb("Batting")
Batting$XBperH <- XBperH(Batting)

}

}
\keyword{XBperH}
\keyword{base}
\keyword{extra}
\keyword{hit}
\keyword{per}
