% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingStats.R
\name{XBperH}
\alias{XBperH}
\title{Calculate extra base per hit}
\usage{
XBperH(dat = NULL)
}
\arguments{
\item{dat}{A data frame you would wish to calculate. If NULL, it will use the appropriate table from
the Lahman package. However, functions will accept custom data frames as well.}
}
\description{
Find the average extra bases per hit for batters with more than zero hits.
Required fields from the batting table are "H", "X2B", "X3B", "HR"."
}
\examples{
\dontrun{
batting_df <- Lahman::Batting
new_df <- XBperH(batting_df)
new_df
}

}
\keyword{XBperH}
\keyword{base}
\keyword{extra}
\keyword{hit}
\keyword{per}

